+++ minijinja/src/vm/context.rs
@@ -1,12 +1,12 @@
 use std::borrow::Cow;
 use std::collections::{BTreeMap, HashSet};
 use std::fmt;
-use std::sync::{Arc, Mutex};
+use std::sync::Arc;
 
 use crate::environment::Environment;
 use crate::error::{Error, ErrorKind};
-use crate::key::{Key, StaticKey};
-use crate::value::{intern, Object, ObjectKind, OwnedValueIterator, StructObject, Value};
+use crate::value::{OwnedValueIterator, Value};
+use crate::vm::closure_object::Closure;
 use crate::vm::loop_object::Loop;
 
 type Locals<'env> = BTreeMap<&'env str, Value>;
@@ -26,46 +26,6 @@ pub(crate) struct LoopState {
     pub(crate) current_recursion_jump: Option<(usize, bool)>,
     pub(crate) iterator: OwnedValueIterator,
     pub(crate) object: Arc<Loop>,
-}
-
-/// Utility to enclose values for macros.
-///
-/// See `closure` on the [`Frame`] for how it's used.
-#[derive(Debug, Default)]
-pub(crate) struct Closure {
-    pub(crate) values: Mutex<BTreeMap<StaticKey, Value>>,
-}
-
-impl fmt::Display for Closure {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let mut m = f.debug_map();
-        for (key, value) in self.values.lock().unwrap().iter() {
-            m.entry(&key, &value);
-        }
-        m.finish()
-    }
-}
-
-impl Object for Closure {
-    fn kind(&self) -> ObjectKind<'_> {
-        ObjectKind::Struct(self)
-    }
-}
-
-impl StructObject for Closure {
-    fn fields(&self) -> Vec<Arc<String>> {
-        self.values
-            .lock()
-            .unwrap()
-            .keys()
-            .filter_map(|x| x.as_str())
-            .map(intern)
-            .collect()
-    }
-
-    fn get_field(&self, name: &str) -> Option<Value> {
-        self.values.lock().unwrap().get(&Key::Str(name)).cloned()
-    }
 }
 
 pub(crate) struct Frame<'env> {
@@ -235,11 +195,7 @@ impl<'env> Context<'env> {
     pub fn store(&mut self, key: &'env str, value: Value) {
         let top = self.stack.last_mut().unwrap();
         if let Some(ref closure) = top.closure {
-            closure
-                .values
-                .lock()
-                .unwrap()
-                .insert(StaticKey::from(key), value.clone());
+            closure.store(key, value.clone());
         }
         top.locals.insert(key, value);
     }
@@ -251,15 +207,8 @@ impl<'env> Context<'env> {
     /// unfortunate downside is that this has to be done with a `Mutex`.
     #[cfg(feature = "macros")]
     pub fn enclose(&mut self, env: &Environment, key: &str) {
-        let value = self.load(env, key);
-        let top = self.stack.last_mut().unwrap();
-        let closure = top
-            .closure
-            .get_or_insert_with(|| Arc::new(Closure::default()));
-        let mut values = closure.values.lock().unwrap();
-        if !values.contains_key(&Key::Str(key)) {
-            values.insert(key.into(), value.unwrap_or(Value::UNDEFINED));
-        }
+        self.closure()
+            .store_if_missing(key, || self.load(env, key).unwrap_or(Value::UNDEFINED));
     }
 
     /// Loads the closure and returns it.
@@ -280,13 +229,13 @@ impl<'env> Context<'env> {
     /// This means that if you override a variable referenced by a macro after
     /// including in the parent template, it will not override the value seen by
     /// the macro.
-    #[cfg(feature = "macros")]
+    #[cfg(feature = "multi_template")]
     pub fn take_closure(&mut self) -> Option<Arc<Closure>> {
         self.stack.last_mut().unwrap().closure.take()
     }
 
     /// Puts the closure back.
-    #[cfg(feature = "macros")]
+    #[cfg(feature = "multi_template")]
     pub fn reset_closure(&mut self, closure: Option<Arc<Closure>>) {
         self.stack.last_mut().unwrap().closure = closure;
     }
